Update taxRate type

+++ search.go
@@ -87,7 +87,7 @@ type CustomOrderSearchReturn struct {
 			TotalPrice      float64 `json:"totalPrice"`
 			CalculatedTaxes []struct {
 				Tax        float64       `json:"tax"`
-				TaxRate    int           `json:"taxRate"`
+				TaxRate    float64       `json:"taxRate"`
 				Price      float64       `json:"price"`
 				Extensions []interface{} `json:"extensions"`
 				ApiAlias   string        `json:"apiAlias"`
@@ -631,7 +631,7 @@ type CustomOrderSearchReturn struct {
 				TotalPrice      float64 `json:"totalPrice"`
 				CalculatedTaxes []struct {
 					Tax        float64       `json:"tax"`
-					TaxRate    int           `json:"taxRate"`
+					TaxRate    float64       `json:"taxRate"`
 					Price      float64       `json:"price"`
 					Extensions []interface{} `json:"extensions"`
 					ApiAlias   string        `json:"apiAlias"`
